requirements/base.txt
# Core Django
Django>=5.0,<6.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9

# Security
dj-database-url>=2.0.0
whitenoise>=6.6.0

# ASGI
uvicorn>=0.22.0

# Utilities
python-decouple>=3.8
gunicorn>=21.2.0

# Production dependencies
dj-database-url>=2.0.0
whitenoise>=6.6.0

# Caching
redis>=5.0.0

# Email
sendgrid>=6.11.0

# Monitoring
sentry-sdk>=1.39.1
